@@ -10,21 +10,7 @@
   },
   "private": true,
   "engines": {
-    "node": "16"
-  },
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "@testing-library/user-event": "^12.6.0",
-    "clsx": "^2.1.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "sonner": "^2.0.3",
-    "starter-restaurant-reservation-front-end": "file:",
-    "web-vitals": "^0.2.4"
+    "node": "18"
   },
   "scripts": {
     "prebuild": "env CI=true react-scripts test --colors",
@@ -65,10 +51,27 @@
       "last 1 safari version"
     ]
   },
+  "dependencies": {
+    "clsx": "^2.1.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router": "^6.30.1",
+    "react-router-dom": "^6.30.1",
+    "react-scripts": "^5.0.1",
+    "sonner": "2.0.3",
+    "starter-restaurant-reservation-front-end": "file:",
+    "web-vitals": "^5.1.0"
+  },
   "devDependencies": {
-    "cross-fetch": "^3.0.6",
-    "jest-puppeteer": "^6.0.0",
+    "@testing-library/jest-dom": "^6.7.0",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "cross-fetch": "^4.1.0",
+    "jest": "^30.0.5",
+    "jest-environment-puppeteer": "^11.0.0",
+    "jest-puppeteer": "^11.0.0",
     "npm-run-all": "^4.1.5",
-    "puppeteer": "^10.4.0"
+    "puppeteer": "^24.16.2",
+    "wait-on": "^8.0.4"
   }
 }
