@@ -2,7 +2,7 @@
  * Express API "Not found" handler.
  */
 function notFound(req, res, next) {
-  next({ status: 404, message: `Path not found: ${req.originalUrl}` });
+	next({ status: 404, message: `Path not found: ${req.originalUrl}` });
 }
 
 module.exports = notFound;
