@@ -4,16 +4,16 @@ const app = require("./app");
 const knex = require("./db/connection");
 
 knex.migrate
-  .latest()
-  .then((migrations) => {
-    console.log("migrations", migrations);
-    app.listen(PORT, listener);
-  })
-  .catch((error) => {
-    console.error(error);
-    knex.destroy();
-  });
+	.latest()
+	.then((migrations) => {
+		console.log("migrations", migrations);
+		app.listen(PORT, listener);
+	})
+	.catch((error) => {
+		console.error(error);
+		knex.destroy();
+	});
 
 function listener() {
-  console.log(`Listening on Port ${PORT}!`);
+	console.log(`Listening on Port ${PORT}!`);
 }
