@@ -1,3 +1,5 @@
 module.exports = {
-  testTimeout: 20000,
+	testTimeout: 20000,
+	rootDir: "../",
+	setupFilesAfterEnv: ["./test/setup.jest.js"],
 };
