@@ -1,4 +1,5 @@
 module.exports = {
-  preset: "jest-puppeteer",
-  testTimeout: 40000,
+	preset: "jest-puppeteer",
+	testTimeout: 60000,
+	verbose: true,
 };
