@@ -1,14 +1,9 @@
+import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 import App from "./App";
-import { BrowserRouter as Router } from "react-router-dom";
-import "@testing-library/jest-dom/extend-expect";
 
 test("renders title", () => {
-  render(
-    <Router>
-      <App />
-    </Router>
-  );
-  const restaurant = screen.getByText(/Menu/i);
-  expect(restaurant).toBeInTheDocument();
+	render(<App />);
+	const restaurant = screen.getByText(/Menu/i);
+	expect(restaurant).toBeInTheDocument();
 });
