@@ -7,8 +7,7 @@ import {
 	default as formatReservationTime,
 } from "./format-reservation-date";
 
-const API_BASE_URL =
-	process.env.REACT_APP_API_BASE_URL || "http://localhost:5001";
+const API_BASE_URL = process.env.BASE_URL || "http://localhost:5001";
 
 /**
  * Defines the default headers for these functions to work with `json-server`
